requirements.txt
streamlit>=1.36.0
openai-whisper>=20231117
moviepy>=1.0.3
imageio-ffmpeg>=0.4.9
ffmpeg-python>=0.2.0
torch>=2.2.0; python_version < "3.13"
